refactor: move pena seasons slice into core application
