+++ backend/src/api/dependencies/use_cases.py
@@ -25,15 +25,15 @@ from core.application.use_cases.manage_pena_membership_usecase import (
 from core.application.use_cases.manage_pena_profile_usecase import (
     ManagePenaProfileUseCase,
 )
+from core.application.use_cases.manage_pena_seasons_usecase import (
+    ManagePenaSeasonsUseCase,
+)
 from core.application.use_cases.register_admin_usecase import RegisterAdminUseCase
 from core.application.use_cases.register_user_usecase import RegisterUserUseCase
 from core.application.use_cases.update_player_profile_usecase import (
     UpdatePlayerProfileUseCase,
 )
 from fastapi import Depends
-from persistence.application.use_cases.manage_pena_seasons_usecase import (
-    ManagePenaSeasonsUseCase,
-)
 from persistence.application.use_cases.manage_season_competition_usecase import (
     ManageSeasonCompetitionUseCase,
 )
