+++ backend/src/api/middleware/exception_mapper.py
@@ -48,6 +48,21 @@ from core.application.use_cases import (
 from core.application.use_cases import (
     PenaSeasonPenaNotFoundError as CoreCompetitionPenaSeasonPenaNotFoundError,
 )
+from core.application.use_cases.manage_pena_seasons_usecase import (
+    InvalidPenaSeasonDataError,
+)
+from core.application.use_cases.manage_pena_seasons_usecase import (
+    PenaSeasonAccessDeniedError as PenaSeasonsAccessDeniedError,
+)
+from core.application.use_cases.manage_pena_seasons_usecase import (
+    PenaSeasonDateOverlapError as PenaSeasonsDateOverlapError,
+)
+from core.application.use_cases.manage_pena_seasons_usecase import (
+    PenaSeasonNotFoundError as PenaSeasonsNotFoundError,
+)
+from core.application.use_cases.manage_pena_seasons_usecase import (
+    PenaSeasonPenaNotFoundError as PenaSeasonsPenaNotFoundError,
+)
 from fastapi import HTTPException, status
 from persistence.application.use_cases import (
     InvalidPenaProfileImageError,
@@ -79,21 +94,6 @@ from persistence.application.use_cases import (
 )
 from persistence.application.use_cases import (
     PenaSeasonPenaNotFoundError as CompetitionPenaSeasonPenaNotFoundError,
-)
-from persistence.application.use_cases.manage_pena_seasons_usecase import (
-    InvalidPenaSeasonDataError,
-)
-from persistence.application.use_cases.manage_pena_seasons_usecase import (
-    PenaSeasonAccessDeniedError as PenaSeasonsAccessDeniedError,
-)
-from persistence.application.use_cases.manage_pena_seasons_usecase import (
-    PenaSeasonDateOverlapError as PenaSeasonsDateOverlapError,
-)
-from persistence.application.use_cases.manage_pena_seasons_usecase import (
-    PenaSeasonNotFoundError as PenaSeasonsNotFoundError,
-)
-from persistence.application.use_cases.manage_pena_seasons_usecase import (
-    PenaSeasonPenaNotFoundError as PenaSeasonsPenaNotFoundError,
 )
 
 ExceptionStatus = tuple[int, str]
